requirements.txt
streamlit>=1.33
pillow
pandas
send2trash
